Add feature toggle to read numeric strings as numeric timestamps
There is an apparent inconsistency in the way Jackson de-serializes numbers
that are shaped as a string into an instant.

 * When not providing a custom date format, quoted numbers are treated as
   epoch milis/nanos
 * When not providing a custom date, quoted numbers are assumed to be handled
   by the pattern.

There is however no way to construct a pattern that handles both ISO dates
and epoch milis/nanos. This feature toggle allow numeric strings to be read
as numeric timestamps.

Fixes https://github.com/FasterXML/jackson-modules-java8/issues/263